Refactor: extract rebate sorting logic to separate module
